nest blueprint - components
