+++ backend/components/__init__.py
@@ -1,0 +1,10 @@
+from flask import Blueprint
+from .home_ext import home_ext
+from .review_ext import review_ext
+from .user_ext import user_ext
+
+
+components = Blueprint("components", __name__)
+components.register_blueprint(home_ext)
+components.register_blueprint(review_ext)
+components.register_blueprint(user_ext)
